Warning: Use v0.0.7 or 0.0.x branch for updates, not main branch
As of this commit, it is recommended to use the latest release
tag(v0.0.7) or the `0.0.x` branch for updates, rather than the main
branch. This is due to ongoing refactoring of the color scheme.

Additionally, the documentation for Vundle installation has been
removed, as it no longer supports specific revisions or branches.

For more information please refer to the installation documentation at
the following link:
https://github.com/projekt0n/github-nvim-theme#installation