breaking-change: Major Changes and Deprecation of some colorschemes
In this commit, I redefined the older highlight groups and removed
support for all plugins listed in the README, including Lualine (which I
redefined in the next commits). Furthermore, all the previous color
palettes have been replaced with Primer.

The configurations `overrides` and `colors` are now deprecated. Please
use `palettes` and `specs` for color overriding as requested in feature
request #176, and use `groups` for overriding the highlight groups.

For more information, please refer to
`:h github-theme.changelog-23042023`.

Additionally, the example has been updated in the README.