fix(autocmds): don't clobber `winhl` nvim option
The user might have their own, custom hlgroup mappings in `winhl` that
we shouldn't be clobbering when they don't conflict with ours.

Also cleanup `autocmds.lua` and ensure that autocmds ALWAYS get
defined/cleared when the colorscheme changes, and make it impossible for
the same autocmd to get defined multiple times.