fix(queries): disable `ecma` query for now
Introduced in the previous commit, this query works with js, jsx, and
tsx, but not ts (i.e. a treesitter query error is thrown when opening a
regular `.ts` typescript file). We will have to figure out a workaround
to get it working with ts as well, but for now we will just disable it.

Issue: #361