refactor: minify colorscheme files in `colors/*.vim`
Refactor and centralize all of the duplicated logic in each of the vim
files in `colors/*.vim` into `require('github-theme').load()`. This
removes duplication and moves all of the logic to Lua (file) where it is
easier to maintain.

Remove the deprecated 'github_dimmed' theme. This has been deprecated
for over a year now, so it is probably a good idea to remove it. Plus it
just wastes space and pollutes cmdline completion with an erroneous
value/theme.